[Dissassemble] - Added scope check.
This is untested. But it looks like this script was changed to get the
parent object each time it downgrades.

So the base class of WoodenFence_1_foundation is WoodenFence_base.

I've added a scope check to make sure the object being spawned is the
correct one. I'll go over this at the weekend.