Replace some forEach with count
The promptRName.sqf file is no longer used. This will be my last commit
with formatting cleanup as I don't want to pollute the commit history.

Count is faster in A2:
https://community.bistudio.com/wiki/Code_Optimisation#forEach_vs_count

And it is safe to use whenever the loop will not be nested inside
another count loop or require _forEachIndex.

UiSleep is more reliable than sleep because it is not slowed down by lag
or disabled simulation:
https://community.bistudio.com/wiki/sleep_vs_uiSleep