Make log more accurate in server_playerSync
"ERROR: Cannot Sync Character as no characterID" will only show when the
player is not in debug. If they are in debug then "Cannot Sync Player.
Position in debug!" will show instead.

Also changed this message to INFO instead of ERROR since it is normal
for it to show when players change clothes (sync is called on loop in
player_spawn2, when gear menu is unloaded and on abort menu onload). If
it is being spammed a lot then the admin knows something is wrong or the
player is changing clothes a lot.