Fix switch weapon addAction with nothing in hands
The addAction click was doing an animation without switching when the
player had a primary or melee on back but nothing in hands.

Also dz_switchWeapon_pistolTime could be undefined in cases where the
player did not switch to a pistol first, since it did not have a default
value.