Faster keydown EH
This eliminates the need for a waitUntil {!isNull findDisplay 46} and
adds the keydown EH as soon as the display is loaded.

It also avoids the bug with onPreloadFinished which causes it to fire
every time the "Receiving.." screen is shown (alt tab, change window
size, etc.). See notes:
https://community.bistudio.com/wiki/onPreloadFinished

The EH is now set only once instead of being reset multiple times.