Basic compatibility fixes for other maps
@skigoggles this fixes most compatibility problems for other maps. To
properly support another map coordinates would need to be added for the
town generator, POIs, infectious water holes and SQL vehicle spawns. If
you get all of that done then it would probably be ok to add a mission
for it with the correct mission.sqm addons, spawn markers, etc. The last
thing would be adding loot positions for non-cherno buildings.