update for #722 vehicles may spawn on top of each other
update for #722 vehicles may spawn on top of each other. This was an edge case where mysql was returning NULL values where it shouldn't. A few servers were seeing this and I was able to randomly reproduce.
The changes to the first loop are to retry the process if this edge case occurs (X or Y is NULL)
The second changes are to prevent anything from being spawned into the same EXACT worldspace (its just a character compare), in case other errors occur.
I have been running this change on my server for months without issue and have not seen the problem again.