added and removed locations
removed duplicates (like klen and others) and thinned out popular areas (Stary/Skalisty). Added new locations for every group. Left the old one in there in case anyone wants to understand the history. Added comments on actual locations - not sure if this is good or not but I figure anyone who can read the github could translate the worldspace anyway.