Fix humanity hit when shooting from vehicle
In player_humanityChange if (_object == player) was returning false if
the player was in a vehicle. This check is not needed anymore anyway
since this script only ever runs on the intended player now (sent with
PVC, not PV). It is not necessary to send the player object either.

The killingBlow setVariable was moved up in player_death because the
same conditions were already checked above.