Fix damage handling of vehicle explosions
This was previously blocking all damage from vehicle explosions.

We only want to block damage from a null source if there is no drivable
vehicle (or drivable vehicle wreck) within a 25m radius. That is the
maximum distance away from a vehicle explosion where the player can
still be damaged by it.

Note I changed "Helicopter" to "Air" because "Air" includes planes.
ParachuteWest is included in both "Helicopter" and "Air":

https://community.bistudio.com/wiki/ArmA_2:_CfgVehicles#Air_Class_Vehicles

Tested with @skigoggles and working well.