Work around long inventory/hitpoints string
Since inventory and hitpoints field in database is of type longtext there is a potential problem here. According to the BI wiki `format` has a limit of 8191 characters. If the player puts many different materials in a large vehicle (for example Ural) this string can become longer and the vehicle does not respawn after server restart.