Remove unnecessary player Respawn event handler
This event handler is redundant since the "Killed" EH already spawns
player_death. I tested and "Killed" also fires when the Respawn button
is used (when legs are broken):

https://github.com/DayZMod/DayZ/blob/Development/SQF/dayz_code/Configs/rscTitles.hpp#L574

https://github.com/DayZMod/DayZ/blob/Development/SQF/dayz_code/medical/setup_functions_med.sqf#L11

It may also prevent _method and _source being properly passed to
player_death due to player_death firing multiple times (depending on
which EH or script spawns it first).

We have been testing with this removed in 1.0.6 for a couple months now
and have not noticed any adverse effects.