Delay player_death 1s if spawned from Killed EH
If player_death is spawned from the Killed EH (no source or method) then
we wait one second to see if sched_medical spawns it with a source and
method.

This should reduce or eliminate cases of method and source being unknown
due to the KilledEH player_death running first. That accounts for
players sometimes not receiving a humanity hit/kill stat for their kill
and body showing unknown cause of death when it shouldn't.