Fix short knockouts for tranquilizer bolt
As reported by Term in Discord testing channel

The knockout is now guaranteed to be 20s-80s for tranq bolts as set in
the damage handler. This change allows r_player_timeout to be pre-set if
desired, instead of always generating a random value based on passed
damage in fnc_usec_damageUnconscious.

Removed redundant setVariable already done in fnc_usec_damageUnconscious