Fix player_breakin.sqf error
Mistake from 416cc5d, getvar on object is undefined the first time the
script runs, need default value.

PVDZ_Server_LogIt is undefined if _proceed = false; and should not send
in that case.

Removed () and [] from PVS due to code filtering in
publicvariableval.txt, and WARNING was already in diag_log message